+++ src/DenseLinAlg/MatrixVector.hpp
@@ -1,0 +1,220 @@
+/*
+ * MatrixVector.hpp
+ *
+ *  Created on: 2015/10/20
+ *      Author: Masakatsu ITO
+ */
+
+#ifndef DENSELINALG_MATRIXVECTOR_HPP_
+#define DENSELINALG_MATRIXVECTOR_HPP_
+
+#include <iostream>
+#include <boost/proto/proto.hpp>
+
+#include <DenseLinAlg/Grammar.hpp>
+
+
+namespace mpl = boost::mpl;
+namespace proto = boost::proto;
+
+
+namespace DenseLinAlg {
+
+
+	// A wrapper for a linear algebraic expression
+	template<typename E> struct ExprWrapper;
+
+	// The above grammar is associated with this domain.
+	struct Domain
+		: proto::domain<proto::generator<ExprWrapper>, ExprGrammar>
+	{};
+
+	// A wrapper template for linear algebraic expressions
+	// including matrices and vectors
+	template<typename ExprType>
+	struct ExprWrapper
+		: proto::extends<ExprType, ExprWrapper<ExprType>, Domain>
+	{
+		/* typedef double result_type; */
+
+		explicit ExprWrapper(const ExprType& e)
+			: proto::extends<ExprType, ExprWrapper<ExprType>, Domain>(e)
+		{}
+	};
+
+
+	// Testing if data in an heap array can be a vector object
+	class Vector {
+		private:
+			int sz;
+			double* data;
+
+	public:
+		template <typename Sig> struct result;
+
+		template <typename This, typename T>
+		struct result< This(T) > { typedef double type; };
+
+		explicit Vector(int sz_ = 1, double iniVal = 0.0) :
+			sz( sz_), data( new double[sz] ) {
+			for (int i = 0; i < sz; i++) data[i] = iniVal;
+			std::cout << "Created" << std::endl;
+		}
+		Vector(const Vector& vec) :
+			sz( vec.sz), data( new double[sz] ) {
+			for (int i = 0; i < sz; i++) data[i] = vec.data[i];
+			std::cout << "Copied! " << std::endl;
+		}
+
+		~Vector() {
+			delete [] data;
+			std::cout << "Deleted" << std::endl;
+		}
+
+		// accessing to an element of this vector
+		double& operator()(int i) { return data[i]; }
+		const double& operator()(int i) const { return data[i]; }
+
+		// assigning the lhs of a vector expression into this vector
+		template<typename Expr>
+		Vector& operator=( const Expr& expr ) {
+			proto::_default<> trans;
+			for(int i=0; i < sz; ++i)
+				data[i] = trans( VecExprGrammar()( expr(i) ) );
+			return *this;
+		}
+
+		// assigning and adding the lhs of a vector expression into this vector
+		template<typename Expr>
+		Vector& operator+=( const Expr& expr ) {
+			proto::_default<> trans;
+			for(int i=0; i < sz; ++i)
+				data[i] += trans( VecExprGrammar()( expr(i) ) );
+			return *this;
+		}
+	};
+
+
+	// Testing if data in an heap array can be a matrix object
+	class Matrix
+	{
+	private:
+		int rowSz, colSz;
+		double* data;
+		double** m;
+
+	public:
+		template <typename Sig> struct result;
+
+		template <typename This, typename T>
+		struct result< This(T,T) > { typedef double type; };
+
+		explicit Matrix(int rowSize = 1, int columnSize =1,
+					double iniVal = 0.0) :
+			rowSz( rowSize), colSz(columnSize),
+			data( new double[rowSz*colSz] ), m( new double*[rowSz])
+		{
+			for (int i = 0; i < rowSz; i++) m[i] = data + i*colSz;
+			for (int ri = 0; ri < rowSz; ri++)
+				for (int ci = 0; ci < colSz; ci++) m[ri][ci] = iniVal;
+			std::cout << "Created" << std::endl;
+		}
+
+		Matrix(const Matrix& mat) :
+			rowSz( mat.rowSz), colSz( mat.colSz),
+			data( new double[rowSz*colSz] ), m( new double*[rowSz])
+		{
+			for (int i = 0; i < rowSz; i++) m[i] = data + i*colSz;
+				for (int ri = 0; ri < rowSz; ri++)
+					for (int ci = 0; ci < colSz; ci++)
+						m[ri][ci] = mat.m[ri][ci];
+				std::cout << "Copied! " << std::endl;
+		}
+
+		~Matrix()
+		{
+			delete [] m;
+			delete [] data;
+			std::cout << "Deleted" << std::endl;
+		}
+
+		int rowSize() const { return rowSz; }
+		int columnSize() const { return colSz; }
+
+		// accesing to a vector element
+		double& operator()(int ri, int ci) { return m[ri][ci]; }
+		const double& operator()(int ri, int ci) const { return m[ri][ci]; }
+
+		// assigning the lhs of a vector expression into this vector
+		template<typename Expr>
+		Matrix& operator=( const Expr& expr ) {
+			proto::_default<> trans;
+			for(int ri=0; ri < rowSz; ri++)
+				for (int ci=0; ci < colSz; ci++ )
+					m[ri][ci] = trans( MatExprGrammar()( expr(ri, ci) ) );
+			return *this;
+		}
+
+		// assigning and adding the lhs of a vector expression into this vector
+		template<typename Expr>
+		Matrix& operator+=( const Expr& expr ) {
+			proto::_default<> trans;
+			for(int ri=0; ri < rowSz; ri++)
+				for (int ci=0; ci < colSz; ci++ )
+					m[ri][ci] += trans( MatExprGrammar()( expr(ri, ci) ) );
+			return *this;
+		}
+	};
+
+	// Lazy function object for evaluating an element of
+	// the resultant vector from the multiplication of
+	// a matrix and vector objects.
+	//
+	// An expression like ( matrix * vector )(index) is transformed
+	// into the loop for calculating the dot product between
+	// the index'th row of the matrix and the vector.
+	struct LazyMatVecMult
+	{
+		Matrix const& m;
+		Vector const& v;
+		int mColSz;
+
+		typedef double result_type;
+		// typedef mpl::int_<1> proto_arity;
+
+		explicit LazyMatVecMult(Matrix const& mat, Vector const& vec) :
+		m( mat), v( vec), mColSz(mat.rowSize()) {}
+
+		LazyMatVecMult(LazyMatVecMult const& lazy) :
+			m(lazy.m), v(lazy.v), mColSz(lazy.mColSz) {}
+
+		result_type operator()(int index) const
+		{
+			result_type elm = 0.0;
+			for (int ci =0;  ci < mColSz; ci++)
+				elm += m(index, ci) * v(ci);
+			return elm;
+		}
+	};
+
+	// Callable transform object to make the lazy functor
+	// a proto exression for lazily evaluationg the multiplication
+	// of a matrix and a vector .
+	struct MatVecMult : proto::callable
+	{
+		typedef proto::terminal< LazyMatVecMult >::type result_type;
+
+		result_type
+		operator()( Matrix const& mat, Vector const& vec) const
+		{
+			return proto::as_expr( LazyMatVecMult(mat, vec) );
+		}
+	};
+
+
+}
+
+
+
+
+#endif /* DENSELINALG_MATRIXVECTOR_HPP_ */
